Adding the blanket library
Added blanket via bower and imported the library when running a test
build.  This imports the library into the require registry, but it may
not be in the right order.  However, importing the library via require
at least ensures that the code is instrumented prior to the qunit start
(this is not the case when referencing the library outside of a module
import)