Resolve "Could not find module" error when used in an Ember CLI Addon
When an addon is built a module is defined with the <addon name> that has a dependency named the name of the
module plus the text of "/index" (ie. `<addon name>/index`).  This dependency does not actually exist, which then causes
the following error to be thrown in the console from the ember-cli/loader.js package:

Error: Could not find module `<addon name>/index` imported from `<addon name>`(…)

This compares the module name created in this situation, <addon name>, against the blanket.options('modulePrefix') and
if they match the/this module should be excluded.